home-agent/requirements.txt
# Home Assistant MCP Client
websockets>=12.0

# MQTT
paho-mqtt>=1.6.0

# Config
pyyaml>=6.0

# Data validation
pydantic>=2.0.0
